requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
scikit-learn>=0.24.2

# Data collection
requests>=2.26.0
beautifulsoup4>=4.9.3
python-dateutil>=2.8.2

# Web scraping
unicodedata2>=14.0.0

# Model training
einops>=0.3.2
tqdm>=4.62.3

# Optimization
pulp>=2.4

# Development tools
jupyter>=1.0.0
black>=21.6b0
flake8>=3.9.0
pytest>=6.2.5